refractored panoptica external app
